requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==25.1
django-extensions==3.2.3
python-decouple==3.8
python-dateutil==2.8.2
openpyxl==3.1.2
reportlab==4.0.9
requests==2.31.0
cryptography==41.0.7
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
celery==5.3.4
dj-database-url==2.1.0
redis==5.0.1
python-dotenv==1.0.0
factory-boy==3.3.0
